+++ src/main/java/au/edu/rmit/sept/webapp/service/UserService.java
@@ -17,7 +17,7 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public Optional<User> findById(Long id) {
+    public Optional<User> findById(int id) {
         return userRepository.findById(id);
     }
